frontend/backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]
SQLAlchemy==2.*
pydantic==2.*
python-dotenv
pymysql
